Updates to remove calls to deprecated functions and to update procedures to fit with the new apis.
Biggest changes in this commit are the updates to the Toolbar module which needed allot of help as wx 2.8 uses a native mac toolbar that didnt quite look right as it was.

This is just an initial commit of a working point in the code in preparation for some more major changes that are to come.