improved drag and drop feedback.
When dragging a file or folder to drop on the editor the cursor is changed to the copy cursor.

When dragging text an image of the text being dragged is drawn.

Would also like to make the carret move with the cursor when dragging text but cant seem to get it to work currently.