Fairly significant performance improvements
1) modified stc event mask to not send events for 
PERFORMED_USER, as this causes modify events to be sent 3 
times per key stroke which is really slow. It now only sends 
one event per key stroke.



2) Only update page title if the value has changed. This was 
done as it is slow to redraw the label on each keystroke. The 
only time it needs to be redrawn is directly after the first 
keystroke from modifying after a save point and when the 
undo stack has been backed up to the last save point.



3) Style sheets are are only parsed once per program launch 
instead of for every document load. The data is cached from 
the initial parse and all subsequent pages use the cached data. 
This noticeably improved the time for opening new files and 
pages, it probably also saves a little memory as only one set of 
style data is kept per program instance instead of one per 
page as previously.